add HttpExceptions if no data

+++ src/mail/mail.service.ts
@@ -1,5 +1,6 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
+
 import { Repository } from 'typeorm';
 import { Mail } from './mail.entity';
 import { MailDTO } from './mail.dto';
@@ -13,12 +14,20 @@ export class MailService {
 
   // Get ALL mails from DB service method
   async getAll() {
-    return (await this.mailRepo.find()) || []; // if no records return empty array
+    const data = await this.mailRepo.find();
+    if (!data) {
+      throw new HttpException('Records not found', HttpStatus.NO_CONTENT);
+    }
+    return data;
   }
 
   // get mail by ID service method
   async get(id: string) {
-    return (await this.mailRepo.findOne(id)) || {};
+    const data = await this.mailRepo.findOne(id);
+    if (!data) {
+      throw new HttpException('Record not found', HttpStatus.NOT_FOUND);
+    }
+    return data;
   }
 
   // add mail service method
@@ -30,11 +39,19 @@ export class MailService {
 
   //  update mail by ID service method
   async update(id: string, data: Partial<MailDTO>) {
+    const dataToUpdate = await this.mailRepo.findOne(id);
+    if (!dataToUpdate) {
+      throw new HttpException(`Record ${id} not found`, HttpStatus.NOT_FOUND);
+    }
     await this.mailRepo.update(id, data); // update data by ID
     return await this.mailRepo.findOne(id); // return updated data from DB
   }
   //  delete mail by ID service method
   async delete(id: string) {
+    const dataToDelete = await this.mailRepo.findOne(id);
+    if (!dataToDelete) {
+      throw new HttpException(`Record ${id} not found`, HttpStatus.NOT_FOUND);
+    }
     await this.mailRepo.delete(id);
     return { deleted: true };
   }
